更新Stock和Event的測試
