+++ Event.h
@@ -25,7 +25,7 @@ private:
     string source; // 事件消息來源
 public:
     Event(unordered_map<string, double> impactWeight, string description, int eventImpact, string source)
-    : impactWeight(impactWeight), description(description), eventImpact(eventImpact), source(source) {} 
+    : impactWeight(impactWeight), description(description), eventImpact(eventImpact), source(source) {} ;
 
 
     /* 可能會用到的成員函數
@@ -35,53 +35,8 @@ public:
     */
 
    // print事件資訊
-   void printEventInfo() const {
-    cout << "Event Information:" << endl;
-    cout << "------------------" << endl;
-    cout << "Description: " << description << endl;
-    cout << "Impact: " << eventImpact << endl;
-    cout << "Source: " << source << endl;
-
-    // 檢查是否有數據
-    if (!impactWeight.empty()) {
-        cout << "Affected Stocks Weights: ";
-        size_t count = 0; // 計數器用於判斷是否是最後一個元素
-        for (const auto& pair : impactWeight) {
-            cout << pair.first << " (" << pair.second << ")";
-            if (++count != impactWeight.size()) cout << ", "; // 如果不是最後一個，添加逗號
-        }
-        cout << endl;
-    } else {
-        cout << "Affected Stocks Weights: No data available yet." << endl;
-    }
-
-    cout << "------------------" << endl;
-   }
+   void printEventDetails() const;
 
 };
-/*
-// ---成員函數---
-void Event::printEventInfo() const {
-    cout << "Event Information:" << endl;
-    cout << "------------------" << endl;
-    cout << "Description: " << description << endl;
-    cout << "Impact: " << eventImpact << endl;
-    cout << "Source: " << source << endl;
 
-    // 檢查是否有數據
-    if (!impactWeight.empty()) {
-        cout << "Affected Stocks Weights: ";
-        size_t count = 0; // 計數器用於判斷是否是最後一個元素
-        for (const auto& pair : impactWeight) {
-            cout << pair.first << " (" << pair.second << ")";
-            if (++count != impactWeight.size()) cout << ", "; // 如果不是最後一個，添加逗號
-        }
-        cout << endl;
-    } else {
-        cout << "Affected Stocks Weights: No data available yet." << endl;
-    }
-
-    cout << "------------------" << endl;
-}
-*/
 #endif
